+++ Assets/Scripts/Health/IDamageAttributes.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamageAttributes
 {
-    public int Damage { get; }
+    public float Damage { get; }
 }
